+++ src/templates/Dashboard.vue
