+++ src/templates/Recipes.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>Recipes</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'Recipes'
+};
+</script>
+
+<style lang="stylus" scoped>
+
+</style>
